+++ src/request.h
@@ -117,10 +117,10 @@ struct BasInfo {
     double peak_max_circ;
 
     bool opt_auto_timer;
-    double opt_auto_timer_seconds;
-    double opt_auto_timer_seconds_old;
+    int opt_auto_timer_seconds;
+    int opt_auto_timer_seconds_old;
     bool opt_auto_timer_started;
-    double opt_auto_timer_seconds_elapsed;
+    int opt_auto_timer_seconds_elapsed;
     enum OptStatus opt_auto_timer_status;
     time_t opt_auto_timer_status_changed;
     bool opt_auto_gas;
